consolidate VagrantBackend unit tests
